fix(creator-outreach): use dev/staging/prod for NODE_ENV, rename emailService to email
- NODE_ENV values: dev, staging, prod (not development/production)
- Rename lib/emailService.js → lib/email.js
- Add ExecPlan to .agent/exec-plans/
