+++ functions/creator-outreach/.env.example
@@ -1,3 +1,3 @@
-NODE_ENV=development
+NODE_ENV=dev
 SENDGRID_API_KEY=SG.YOUR-SENDGRID-API-KEY
 SENTRY_DSN=https://YOUR-DSN@o0000000000.ingest.us.sentry.io/0000000000
