+++ functions/creator-outreach/lib/email.js
