+++ functions/creator-outreach/lib/sentry.js
@@ -2,7 +2,7 @@ const Sentry = require("@sentry/node");
 
 Sentry.init({
   dsn: process.env.SENTRY_DSN,
-  environment: process.env.NODE_ENV || "development",
+  environment: process.env.NODE_ENV || "dev",
 });
 
 module.exports = Sentry;
